asset arithmetic: three unchecked integer ranges -- #310
Three unrelated sites, one shape: an index range validated with arithmetic that
could not represent the answer.

collision_frame's overlap narrowphase read cell_triangle_indices[i] with no
bound on i. The nearby check bounds the LENGTH of cell_offsets and the decoder
bounds only the element COUNTS, so nothing requires the offsets themselves to
be monotonic or within cell_triangle_indices. The `tri` check immediately below
already guarded the second hop. Both sibling walks of this same CSR structure
carry the missing line -- collision_surface_sampling.cpp and
behavior_plugin_adapter.cpp -- so this was an omission, not a contract the
other two were being paranoid about.

The surfel density range was validated as `first_surfel + surfel_count >
surfels.size()` in uint32 at THREE sites -- decoder, valid(), and the consumer
that then indexes surfels[first_surfel + i]. With first_surfel = 0xFFFFFFFF and
surfel_count = 1 the sum is 0, so the test is false and the entry is accepted.
Because all three shared the expression there was no defence in depth: a
wrapped range passed every gate on its way to the subscript. Widened to 64-bit.

The raw-F32 scalar field computed sample count and byte count in uint32.
65536 x 16384 x 1 are each inside their declared range, but the product is 2^30
and its byte count 2^32 wraps to ZERO -- so the guard read `size() != 0`, an
EMPTY file passed, and the compiler allocated 4 GiB of zeros and published a
65536x16384 resident texture plus a full mip pyramid. count = 2^30+1 wraps to 4
instead, so a four-byte file bought the same field. ScalarFieldData::count()
shares the 32-bit product, so valid() agreed with the wrapped arithmetic rather
than catching it. Both products are now 64-bit, and a shape that cannot be
expressed in the 32-bit counts used downstream is refused rather than
truncated into one that can.

Note the declared ParamDecl min/max do not bound any of this: they are
documented as slider range only and params.get<uint32_t> does not clamp, so a
hand-edited assets.graph.json or an ABI set_param arrives unfiltered.

ctest 315, green.

+++ src/engine/assets/renderable/renderable_compilers.cpp
@@ -898,8 +898,13 @@ namespace wz::engine::assets::internal
             const TerrainVisualProxyChunkRecord& chunk,
             const TerrainVisualProxySurfelDensityLevel& level)
         {
+            // Third instance of the same wrapping uint32 sum (issue #310,
+            // A4-C27), and the one that goes on to index
+            // chunk.surfels[first_surfel + i]. All three sites carried it, so
+            // there was no defence in depth: a wrapped range passed every gate.
             if (!level.valid()
-                || level.first_surfel + level.surfel_count
+                || static_cast<std::uint64_t>(level.first_surfel)
+                       + static_cast<std::uint64_t>(level.surfel_count)
                     > chunk.surfels.size())
             {
                 return;
